+++ src/components/Header/Header.jsx
@@ -1,11 +1,8 @@
 import { useState } from 'react';
-import { FaShoppingCart, FaUser } from 'react-icons/fa';
+import { FaUser } from 'react-icons/fa';
 import { Link } from 'react-router-dom';
 
-import FavoriteIcon from '@mui/icons-material/Favorite';
 import MenuIcon from '@mui/icons-material/Menu';
-import { List } from '@mui/material';
-import Badge from '@mui/material/Badge';
 import IconButton from '@mui/material/IconButton';
 import Menu from '@mui/material/Menu';
 import MenuItem from '@mui/material/MenuItem';
@@ -13,8 +10,8 @@ import SwipeableDrawer from '@mui/material/SwipeableDrawer';
 import Toolbar from '@mui/material/Toolbar';
 
 import { route } from '@/constants/route.js';
-import styled from '@emotion/styled';
 import { NavigateMenu } from '../NavigateMenu/NavigateMenu.jsx';
+import { UserMenu } from '../UserMenu/UserMenu.jsx';
 import { AppBarStyled, AppContainerStyled } from './Header.styled.js';
 
 // import SearchBar from './searchBar.jsx';
@@ -22,17 +19,6 @@ import Logo from '../../assets/images/logo.svg';
 
 // Mock user data
 const user = { isAuth: true, roles: ['user', 'chef', 'admin', ''] };
-
-const StyledBadge = styled(Badge)(({ theme }) => ({
-  '& .MuiBadge-badge': {
-    right: -3,
-    top: 13,
-    border: `2px solid ${theme.palette.background.paper}`,
-    padding: '0 4px',
-    backgroundColor: 'orange',
-    color: 'black',
-  },
-}));
 
 const Header = () => {
   const [anchorEl, setAnchorEl] = useState(null);
@@ -67,47 +53,33 @@ const Header = () => {
         </Link>
 
         <Toolbar>
-          <IconButton onClick={handleMenuOpen}>
-            <FaUser />
-          </IconButton>
-          <Menu
-            anchorEl={anchorEl}
-            open={Boolean(anchorEl)}
-            onClose={handleMenuClose}
-          >
-            {!user.isAuth && (
-              <MenuItem onClick={handleMenuClose}>
-                <Link to={route.SIGN_UP}>Sign up</Link>
-              </MenuItem>
-            )}
-            {!user.isAuth && (
-              <MenuItem onClick={handleMenuClose}>
-                <Link to={route.SIGN_IN}>Login</Link>
-              </MenuItem>
-            )}
-            {user.isAuth && (
-              <MenuItem onClick={handleMenuClose}>
-                <Link to={route.SIGN_IN}>Logout</Link>
-              </MenuItem>
-            )}
-          </Menu>
-          {user.isAuth && (
-            <List>
-              <IconButton>
-                <StyledBadge badgeContent={4} color="success">
-                  <FavoriteIcon />
-                </StyledBadge>
+          {!user.isAuth && (
+            <>
+              <IconButton onClick={handleMenuOpen}>
+                <FaUser />
               </IconButton>
-              <IconButton>
-                <StyledBadge badgeContent={4} color="success">
-                  <FaShoppingCart />
-                </StyledBadge>
-              </IconButton>
-            </List>
+              <Menu
+                anchorEl={anchorEl}
+                open={Boolean(anchorEl)}
+                onClose={handleMenuClose}
+              >
+                {!user.isAuth && (
+                  <MenuItem onClick={handleMenuClose}>
+                    <Link to={route.SIGN_UP}>Sign up</Link>
+                  </MenuItem>
+                )}
+                {!user.isAuth && (
+                  <MenuItem onClick={handleMenuClose}>
+                    <Link to={route.SIGN_IN}>Sign in</Link>
+                  </MenuItem>
+                )}
+              </Menu>
+            </>
           )}
+          {user.isAuth && <UserMenu />}
 
           <IconButton aria-label="menu" onClick={toggleDrawer(true)}>
-            <MenuIcon />
+            <MenuIcon sx={{ width: 35, height: 35 }} />
           </IconButton>
         </Toolbar>
       </AppContainerStyled>
